zsh: Use ZDOTDIR
Change the zsh file structure to be consistent with what zsh expects to
find in the ZDOTDIR directory.

Environment variables are moved to .zshenv, and interactive shell
commands are moved to .zshrc.

The ZDOTDIR variable itself still needs to be set somewhere else, since
it's what's used to find these files in the first place.